Improved CI tests with REDIS_PASSWORD test

+++ init.d/60-redis.sh
@@ -12,10 +12,12 @@ echo "NOTICE: Initializing settings"
 # Setup the password
 if [ -n "$REDIS_PASSWORD" ]; then
 	# Enable ACL file
-	sed -ri "s!^#?(aclfile)\s*=\s*\S+.*!\1 = /etc/redis/users.acl!" /etc/redis.conf
-	grep -F "aclfile = /etc/redis/users.acl" /etc/redis.conf
+	sed -ri "s!^#?\s*(aclfile)\s+\S+.*!\1 /etc/redis/users.acl!" /etc/redis.conf
+	grep -q -E "^aclfile /etc/redis/users.acl" /etc/redis.conf
 	# Setup default user
-	echo "user default on +@all ~* >$REDIS_PASSWORD" > /etc/redis/users.acl
+	if [ ! -s /etc/redis/users.acl ]; then
+		echo "user default on +@all ~* >$REDIS_PASSWORD" > /etc/redis/users.acl
+	fi
 	# Fixup perms
 	chmod 0640 /etc/redis/users.acl
 	chown root:redis /etc/redis/users.acl
